Big Update

+++ src/controllers/TransactionController.js
@@ -1,11 +1,19 @@
 import { PrismaClient } from '@prisma/client';
-import { StatisticsService } from '../services/StatisticsService'
+import { StatisticsService } from '../services/StatisticsService';
 const prisma = new PrismaClient();
 
 export default {
     async createTransaction(request, reply) {
         try {
             const { description, price, type } = request.body;
+
+            if (isNaN(Number(price))) {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid price format.', status: 'BAD_REQUEST' } });
+            }
+
+            if (typeof type !== 'boolean') {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid type format.', status: 'BAD_REQUEST' } });
+            }
 
             const transaction = await prisma.Transactions.create({
                 data: {
@@ -17,25 +25,27 @@ export default {
 
             return reply.send(transaction);
         } catch (error) {
-            return reply.code(500).send(error);
+            return reply.code(500).send({ error: { code: 500, message: 'Internal Server Error.', status: 'INTERNAL_SERVER_ERROR' } });
         }
     },
 
     async findTransactions(request, reply) {
         try {
-            const take = 10;
-            const skip = Number(request.query?.skip) || 0;
+            const pageNum = Number(request.params?.pageNum) || 1;
+
+            const maxItems = 10;
+            const offset = (pageNum - 1) * maxItems;
 
             const [ transactions, total ] = await prisma.$transaction([
-                prisma.Transactions.findMany({ take, skip }),
+                prisma.Transactions.findMany({ take: maxItems, skip: offset }),
                 prisma.Transactions.count()
             ])
 
-            const totalPages = Math.ceil(total / take)
+            const totalPages = Math.ceil(total / maxItems)
 
             return reply.send({ total, totalPages, transactions });
         } catch (error) {
-            return reply.code(500).send(error);
+            return reply.code(500).send({ error: { code: 500, message: 'Internal Server Error.', status: 'INTERNAL_SERVER_ERROR' } });
         }
     },
 
@@ -47,7 +57,7 @@ export default {
 
             return reply.send(statisticsService);
         } catch (error) {
-            return reply.code(500).send(error);
+            return reply.code(500).send({ error: { code: 500, message: 'Internal Server Error.', status: 'INTERNAL_SERVER_ERROR' } });
         }
     },
 
@@ -55,10 +65,22 @@ export default {
         try {
             const { id, description, price, type } = request.body;
 
+            if (isNaN(Number(id))) {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid id format.', status: 'BAD_REQUEST' } });
+            }
+
+            if (isNaN(Number(price))) {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid price format.', status: 'BAD_REQUEST' } });
+            }
+
+            if (typeof type !== 'boolean') {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid type format.', status: 'BAD_REQUEST' } });
+            }
+
             let transaction = await prisma.Transactions.findUnique({ where: { id: Number(id) } });
 
             if (!transaction) {
-                return reply.code(400).send({ error: { code: 400, message: 'Transaction not found.', status: 'FAILED_PRECONDITION' } });
+                return reply.code(404).send({ error: { code: 404, message: 'Transaction not found.', status: 'NOT_FOUND' } });
             }
 
             transaction = await prisma.Transactions.update({
@@ -68,7 +90,7 @@ export default {
 
             return reply.send(transaction);
         } catch (error) {
-            return reply.code(500).send(error);
+            return reply.code(500).send({ error: { code: 500, message: 'Internal Server Error.', status: 'INTERNAL_SERVER_ERROR' } });
         }
     },
 
@@ -76,17 +98,21 @@ export default {
         try {
             const { id } = request.params;
 
-            let transaction = await prisma.Transactions.findUnique({ where: { id: Number(id) } });
+            if (isNaN(Number(id))) {
+                return reply.code(400).send({ error: { code: 400, message: 'Invalid ID format.', status: 'BAD_REQUEST' } });
+            }
+
+            const transaction = await prisma.Transactions.findUnique({ where: { id: Number(id) } });
 
             if (!transaction) {
-                return reply.code(400).send({ error: { code: 400, message: 'Transaction not found.', status: 'FAILED_PRECONDITION' } });
+                return reply.code(404).send({ error: { code: 404, message: 'Transaction not found.', status: 'NOT_FOUND' } });
             }
 
             await prisma.Transactions.delete({ where: { id: Number(id) } });
 
             return reply.send('Transaction deleted successfully.');
         } catch (error) {
-            return reply.code(500).send(error);
+            return reply.code(500).send({ error: { code: 500, message: 'Internal Server Error.', status: 'INTERNAL_SERVER_ERROR' } });
         }
     },
 };
